docxer/pyproject.toml
[project]
name = "docxer"
version = "0.1.0"
description = "Convert GitHub-flavored Markdown to Microsoft Word documents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentschema>=0.1.2a1",
    "mistune>=3.2.0",
    "python-docx>=1.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "azure-ai-projects==2.0.0b3",
    "azure-identity>=1.25.1",
    "python-dotenv>=1.2.1",
    "openai>=2.15.0",
    "aiohttp>=3.13.3",
]

[project.scripts]
docxer = "docxer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pydantic>=2.12.5",
    "types-pyyaml>=6.0.12.20250915",
]
